fix test imports by stubbing deps early
