+++ docs/ARCHITECTURE_OVERVIEW.md
@@ -96,8 +96,7 @@ adapter layer.
 
 Sanitized events are appended to a lightweight ledger along with their
 Redpanda offsets. The ledger can be queried via the `/ledger/events` API and
-used with `PersistentGraph.replay_from_ledger()` to rebuild state from any
-offset.
+used with `ume.replay.replay_from_ledger()` to rebuild state from any offset.
 
 ## Policy DSL Flow
